+++ integration/src/integration-logger.js
@@ -1,0 +1,9 @@
+export class IntegrationLogger {
+  constructor() {
+    this.entries = [];
+  }
+
+  log(entry) {
+    this.entries.push(entry);
+  }
+}